Update 3ware check to not alert on VERIFYING disks
Right now this 3ware check has the ability to allow for VERIFYING to be
treated as an acceptable state for arrays.  However, this state is never
acceptable for individual drives. Plumb this logic through to the
test_drives function - unless warn_true == True, don't fire events for:
VERIFYING, INITIALIZING, or REBUILDING