curl_redirect
Instead of implementing the redirect feature for each plugins separately, this functionality has been moved to lib/HariSekhonUtils.pm; the plugins now are changed to use only curl_redirect.

Redirect feature: if a plugin encounters the response code 307, it exits directly with status ok, and message 'Standby with active at $active'.
First advantage: this is actually the correct state (instead of following the redirect)
Second advantage: it works in conjunction with Kerberos